LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-expected.txt


FAIL Popover focus test: default behavior - popover is not focused assert_false: Opening a modal dialog should hide the popover expected false got true
FAIL Popover button click focus test: default behavior - popover is not focused assert_false: popover should start out hidden expected false got true
PASS Popover corner cases test: default behavior - popover is not focused
FAIL Popover focus test: autofocus popover assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
FAIL Popover button click focus test: autofocus popover assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
PASS Popover corner cases test: autofocus popover
FAIL Popover focus test: autofocus empty popover assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
FAIL Popover button click focus test: autofocus empty popover assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
PASS Popover corner cases test: autofocus empty popover
FAIL Popover focus test: autofocus popover with button assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
FAIL Popover button click focus test: autofocus popover with button assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
PASS Popover corner cases test: autofocus popover with button
FAIL Popover focus test: autofocus child assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
FAIL Popover button click focus test: autofocus child assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
PASS Popover corner cases test: autofocus child
FAIL Popover focus test: autofocus on tabindex=0 element assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
FAIL Popover button click focus test: autofocus on tabindex=0 element assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
PASS Popover corner cases test: autofocus on tabindex=0 element
FAIL Popover focus test: autofocus multiple children assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
FAIL Popover button click focus test: autofocus multiple children assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
PASS Popover corner cases test: autofocus multiple children
FAIL Popover focus test: autofocus popover and multiple autofocus children assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
FAIL Popover button click focus test: autofocus popover and multiple autofocus children assert_equals: expected Element node <button id="priorFocus"></button> but got Element node <body>

<div popover="" data-test="autofocus popover" aut...
PASS Popover corner cases test: autofocus popover and multiple autofocus children
